+++ client/phpTest/demo/baseInclude.php
@@ -1,0 +1,6 @@
+<?php
+
+use Hvy\CraneDemo\ConnectionDB;
+
+$dbh = new ConnectionDB();
+$dbh->openDbConnection();